pyproject.toml
[project]
name = "wallet-balance-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.2",
    "httpx>=0.28.1",
    "prometheus-client>=0.24.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
